requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
requests>=2.32.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
Pillow>=10.0.0
